# openapi.yaml
...
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5dc0d95e9ef7a1181f8e1322
  title: khachsan
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /room/:
    post:
      x-skaffolder-id: 5dc0ee879ef7a1181f8e1463
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5dc0e8119ef7a1181f8e13f2
      x-skaffolder-resource: room
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - room
      parameters:
        - name: obj
          x-skaffolder-type: room
          in: path
          description: Object to insert
          required: true
          schema:
            type: room
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/room"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                employessid:
                  type: integer
                idservise:
                  type: integer
                note:
                  type: string
                roomname:
                  type: string
                roomstauts:
                  type: string
                roomtype:
                  type: string
                salary:
                  type: number
    get:
      x-skaffolder-id: 5dc0ee6b9ef7a1181f8e145e
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5dc0e8119ef7a1181f8e13f2
      x-skaffolder-resource: room
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: ARRAY OF room
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - room
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/room"
  /room/{id}:
    delete:
      x-skaffolder-id: 5dc0ee8e9ef7a1181f8e1465
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5dc0e8119ef7a1181f8e13f2
      x-skaffolder-resource: room
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - room
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id room
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5dc0ee7d9ef7a1181f8e1461
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5dc0e8119ef7a1181f8e13f2
      x-skaffolder-resource: room
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: room
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - room
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id room
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/room"
    post:
      x-skaffolder-id: 5dc0ee759ef7a1181f8e145f
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5dc0e8119ef7a1181f8e13f2
      x-skaffolder-resource: room
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: room
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - room
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id room
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/room"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/room"
  /room/findByroomname/{key}:
    get:
      x-skaffolder-id: 5dc0ee509ef7a1181f8e145c
      x-skaffolder-name: findByroomname
      x-skaffolder-id-resource: 5dc0e8119ef7a1181f8e13f2
      x-skaffolder-resource: room
      x-skaffolder-crudAction: findByroomname
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findByroomname
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findByroomname/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findByroomname
      tags:
        - room
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of the resource roomname to search
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/room"
    
  #-----  servise  -----
  /servise/:
    post:
      x-skaffolder-id: 5dc0f0079ef7a1181f8e148d
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5dc0eab89ef7a1181f8e1411
      x-skaffolder-resource: servise
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - servise
      parameters:
        - name: obj
          x-skaffolder-type: servise
          in: path
          description: Object to insert
          required: true
          schema:
            type: servise
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/servise"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nameservise:
                  type: string
                salary:
                  type: number
    get:
      x-skaffolder-id: 5dc0f0039ef7a1181f8e148c
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5dc0eab89ef7a1181f8e1411
      x-skaffolder-resource: servise
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: ARRAY OF servise
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - servise
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/servise"
  /servise/{id}:
    delete:
      x-skaffolder-id: 5dc0f0129ef7a1181f8e148f
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5dc0eab89ef7a1181f8e1411
      x-skaffolder-resource: servise
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - servise
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id servise
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5dc0effc9ef7a1181f8e148a
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5dc0eab89ef7a1181f8e1411
      x-skaffolder-resource: servise
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: servise
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - servise
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id servise
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/servise"
    post:
      x-skaffolder-id: 5dc0eff89ef7a1181f8e1488
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5dc0eab89ef7a1181f8e1411
      x-skaffolder-resource: servise
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: servise
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - servise
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id servise
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/servise"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/servise"
    
    
# JWT security
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    room:
      x-skaffolder-id: 5dc0e8119ef7a1181f8e13f2
      x-skaffolder-id-db: 5dc0d95e9ef7a1181f8e1323
      x-skaffolder-id-entity: 5dc0e8119ef7a1181f8e13eb
      x-skaffolder-url: /room
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        employessid:
          type: integer
          x-skaffolder-id-attr: 5dc0e85f9ef7a1181f8e13f4
          x-skaffolder-type: Integer
        idservise:
          type: integer
          x-skaffolder-id-attr: 5dc0e8119ef7a1181f8e13f1
          x-skaffolder-type: Integer
        note:
          type: string
          x-skaffolder-id-attr: 5dc0e8119ef7a1181f8e13ef
          x-skaffolder-type: String
        roomname:
          type: string
          x-skaffolder-id-attr: 5dc0e8119ef7a1181f8e13ec
          x-skaffolder-type: String
        roomstauts:
          type: string
          x-skaffolder-id-attr: 5dc0e8119ef7a1181f8e13f0
          x-skaffolder-type: String
        roomtype:
          type: string
          x-skaffolder-id-attr: 5dc0e8119ef7a1181f8e13ed
          x-skaffolder-type: String
        salary:
          type: number
          x-skaffolder-id-attr: 5dc0e8119ef7a1181f8e13ee
          x-skaffolder-type: Decimal
      x-skaffolder-relations: 
        coderoom:
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 5dc0e8119ef7a1181f8e13eb
          x-skaffolder-ent2: 5dc0ea4b9ef7a1181f8e1401
        roomid:
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 5dc0e8119ef7a1181f8e13eb
          x-skaffolder-ent2: 5dc0e8d59ef7a1181f8e13f6
      required:
        - _id
    servise:
      x-skaffolder-id: 5dc0eab89ef7a1181f8e1411
      x-skaffolder-id-db: 5dc0d95e9ef7a1181f8e1323
      x-skaffolder-id-entity: 5dc0eab79ef7a1181f8e140e
      x-skaffolder-url: /servise
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        nameservise:
          type: string
          x-skaffolder-id-attr: 5dc0eab79ef7a1181f8e140f
          x-skaffolder-type: String
        salary:
          type: number
          x-skaffolder-id-attr: 5dc0eab79ef7a1181f8e1410
          x-skaffolder-type: Number
      x-skaffolder-relations: 
        idservise:
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 5dc0eab79ef7a1181f8e140e
          x-skaffolder-ent2: 5dc0e8119ef7a1181f8e13eb
      required:
        - _id
      
    
    
  # Databases
...